+++ src/main/java/learn/ang/umsback/model/User.java
@@ -31,4 +31,10 @@ public class User {
     @Enumerated(EnumType.STRING)
     @Column(name = "role", nullable = false)
     private Role role;
+
+    @Transient
+    private String accessToken;
+
+    @Transient
+    private String refreshTokenId;
 }