+++ src/main/java/learn/ang/umsback/repository/JwtRefreshTokenRepository.java
@@ -1,0 +1,7 @@
+package learn.ang.umsback.repository;
+
+import learn.ang.umsback.model.JwtRefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JwtRefreshTokenRepository extends JpaRepository<JwtRefreshToken, String> {
+}