+++ src/main/java/learn/ang/umsback/service/JwtRefreshTokenService.java
@@ -1,0 +1,11 @@
+package learn.ang.umsback.service;
+
+import learn.ang.umsback.model.JwtRefreshToken;
+import learn.ang.umsback.model.User;
+
+public interface JwtRefreshTokenService {
+
+    JwtRefreshToken createRefreshToken(Long userId);
+
+    User generateAccessTokenFromRefreshToken(String refreshTokenId);
+}